+++ programs/turf_vault/src/instructions/mod.rs
@@ -13,6 +13,7 @@ pub mod update_signers;
 pub mod mint_entry_token;
 pub mod enter_contest_with_token;
 pub mod enter_contest_direct_with_token;
+pub mod create_season;
 
 pub use initialize::*;
 pub use create_user_account::*;
@@ -29,3 +30,4 @@ pub use update_signers::*;
 pub use mint_entry_token::*;
 pub use enter_contest_with_token::*;
 pub use enter_contest_direct_with_token::*;
+pub use create_season::*;